kubernetes/requirements.txt
Jinja2==2.8
kubernetes==2.0.0
MarkupSafe==0.23
PyYAML==3.12
